gyak_pattern/requirements.txt
python-docx
striprtf
